# Conflicts:
#	engine/gameobject/src/gameobject/gameobject.cpp
#	engine/gameobject/src/gameobject/gameobject_props.cpp
#	engine/gameobject/src/gameobject/gameobject_props.h
#	engine/gameobject/src/gameobject/res_collection.cpp
#	engine/gameobject/src/gameobject/res_prototype.cpp
#	engine/gamesys/src/gamesys/components/comp_label.cpp
#	engine/gamesys/src/gamesys/components/comp_model.cpp
#	engine/gamesys/src/gamesys/components/comp_model.h
#	engine/gamesys/src/gamesys/components/comp_private.cpp
#	engine/gamesys/src/gamesys/components/comp_sprite.cpp
#	engine/gamesys/src/gamesys/components/comp_tilegrid.cpp
#	engine/gamesys/src/gamesys/gamesys_private.h
#	engine/gamesys/src/gamesys/scripts/script_label.cpp
#	engine/gamesys/src/gamesys/scripts/script_model.cpp
#	engine/gamesys/src/gamesys/scripts/script_sprite.cpp
#	engine/resource/src/resource.h

+++ engine/dlib/src/dlib/connection_pool.cpp
@@ -311,7 +311,7 @@ namespace dmConnectionPool
         }
 
         uint64_t handshakestart = dmTime::GetTime();
-        if( timeout > 0 && (handshakestart - connectstart) > timeout )
+        if( timeout > 0 && (handshakestart - connectstart) > (uint64_t)timeout )
         {
             r = RESULT_SOCKET_ERROR;
             dmSocket::Delete(c->m_Socket);
